+++ tests/test_cli.py
@@ -115,3 +115,22 @@ def test_receive_tokenv1(mint):
 
     print("RECEIVE")
     print(result.output)
+
+
+@pytest.mark.asyncio()
+def test_nostr_send(mint):
+    runner = CliRunner()
+    result = runner.invoke(
+        cli,
+        [
+            *cli_prefix,
+            "send",
+            "1",
+            "-n",
+            "aafa164a8ab54a6b6c67bbac98a5d5aec7ea4075af8928a11478ab9d74aec4ca",
+            "-y",
+        ],
+    )
+
+    print("NOSTR_SEND")
+    print(result.output)
